requirements.txt
openai
autogluon
pyyaml
